skillsync-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
google-generativeai==0.5.4
Pillow==10.3.0
python-dotenv==1.0.1
librosa>=0.10.0
soundfile>=0.12.0
langdetect>=1.0.9
pydub>=0.25.1
